fix: cast day column to int in SEIR model to match observed data type
  - Also fix issue with non-existent diagnostics attribute.

  Simulation was outputting day as f64 while observed data has i64, causing
  join failures in Calabaria target evaluation.

+++ examples/simulation-workflow/models/seir.py
@@ -71,7 +71,7 @@ class StochasticSEIR(BaseModel):
         # Simulation loop
         for t in times:
             if int(t) == t:  # Record daily
-                result_times.append(float(t))
+                result_times.append(int(t))  # Store as int for day column
                 result_I.append(I)
 
             # Rates
